+++ R/with_ggtrace.R
@@ -7,12 +7,16 @@
 #'
 #' @param x A ggplot object whose evaluation triggers the trace as specified by the `...`
 #' @inheritDotParams ggtrace
+#' @param return_value Whether the function should return the output of triggered traces
+#'   ("tracedump"), or the resulting graphical object from evaluating the ggplot ("gtable"),
+#'   or "both", which returns the tracedump but also renders the resulting plot as a
+#'   side effect. Defaults to "tracedump".
 #'
 #' @note To force evaluation of `x`, `ggeval_silent(x)` is called internally.
 #'
 #' @seealso [ggtrace()], [ggeval_silent()]
 #'
-#' @return A list
+#' @return A list or gtable object of class `<ggtrace_highjacked>`
 #' @export
 #'
 #' @examples
@@ -36,7 +40,7 @@
 #'
 #' identical(first_tracedump, second_tracedump)
 #'
-with_ggtrace <- function(x, ...) {
+with_ggtrace <- function(x, ..., return_value = c("tracedump", "gtable", "both")) {
   if (!inherits(x, "ggplot")) { rlang::abort("`x` must be a ggplot object") }
   suppressMessages({
     prev_silent_opt <- getOption("ggtrace.suppressMessages")
@@ -48,21 +52,28 @@ with_ggtrace <- function(x, ...) {
     clear_global_ggtrace()
 
     ggtrace(...)
-    ggeval_silent(x)
+    fig <- ggeval_silent(x)
     gguntrace(...)
 
-    out <- global_ggtrace()
-    if (length(out) == 1L) { out <- out[[1]] }
+    dump <- global_ggtrace()
+    if (length(dump) == 1L) { dump <- dump[[1]] }
 
     set_last_ggtrace(prev_last)
     global_ggtrace_state(prev_global_state)
     set_global_ggtrace(prev_global)
     options("ggtrace.suppressMessages" = prev_silent_opt)
   })
-  if (is.null(out)) {
-    rlang::abort("No function to capture - did the ggplot call the method?")
+  if (is.null(dump)) {
+    rlang::abort("Trace was not triggered - did the plot call the method?")
   }
-  out
+
+  switch(
+    match.arg(return_value),
+    "tracedump" = dump,
+    "gtable"    = invisible(fig),
+    "both" = {
+      ggdraw_silent(fig)
+      dump
+    }
+  )
 }
-
-